Add star finder utils in new module
